pulling combustor out.
